Warn if directories have files in them
This is the simplest solution we have currently, just to leave those
directories.  A better solution would be to check if there are supposed to
be other files in their (using the package index), then only warn if there
are not.

+++ python/xpkg/core.py
@@ -455,7 +455,10 @@ class Environment(object):
                 # link file itself is present.
                 if os.path.lexists(full_path):
                     if os.path.isdir(full_path):
-                        os.rmdir(full_path)
+                        if len(os.listdir(full_path)) == 0:
+                            os.rmdir(full_path)
+                        else:
+                            print 'WARNING: not removing dir, has files:',full_path
                     else:
                         os.remove(full_path)
                 else:
